fix: cyrogenic should working properly on players
