+++ src/main/java/dev/galacticraft/mod/mixin/client/LivingEntityRendererMixin.java
@@ -24,7 +24,7 @@ package dev.galacticraft.mod.mixin.client;
 
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.math.Axis;
-import dev.galacticraft.mod.accessor.LivingEntityAccessor;
+import dev.galacticraft.mod.accessor.CryogenicAccessor;
 import dev.galacticraft.mod.content.entity.RocketEntity;
 import net.minecraft.client.model.EntityModel;
 import net.minecraft.client.renderer.MultiBufferSource;
@@ -55,7 +55,7 @@ public abstract class LivingEntityRendererMixin {
 
     @Redirect(method = "render(Lnet/minecraft/world/entity/LivingEntity;FFLcom/mojang/blaze3d/vertex/PoseStack;Lnet/minecraft/client/renderer/MultiBufferSource;I)V", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/LivingEntity;hasPose(Lnet/minecraft/world/entity/Pose;)Z"))
     private boolean gc$hasSleepPose(LivingEntity instance, Pose pose) {
-        if (((LivingEntityAccessor)instance).isInCryoSleep())
+        if (instance instanceof CryogenicAccessor player && player.isInCryoSleep())
             return false;
         return instance.hasPose(pose);
     }
@@ -74,7 +74,7 @@ public abstract class LivingEntityRendererMixin {
 
     @Inject(method = "setupRotations", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/LivingEntity;getBedOrientation()Lnet/minecraft/core/Direction;"), cancellable = true)
     private void galacticraft$renderCryoChamberPos(LivingEntity livingEntity, PoseStack poseStack, float f, float g, float h, CallbackInfo ci) {
-        if (livingEntity instanceof LivingEntityAccessor livingEntityAccessor && livingEntityAccessor.isInCryoSleep()) {
+        if (livingEntity instanceof CryogenicAccessor livingEntityAccessor && livingEntityAccessor.isInCryoSleep()) {
             Direction direction = livingEntity.getBedOrientation();
             float j = direction != null ? sleepDirectionToRotationCryo(direction) : g;
             poseStack.mulPose(Axis.YP.rotationDegrees(j));
